Change external-mdns to a daemonset
This needs to be a daemonset because the node it's running on can't
resolve the hosts, probably because it sends out the mdns messages, but
never gets them itself. We need this running on at least 2 k8s hosts so
that both hosts running it can get the mdns messages and resolve the
hostnames.